+++ lib/committee.rb
@@ -6,7 +6,6 @@ require_relative "committee/errors"
 require_relative "committee/string_params_coercer"
 require_relative "committee/parameter_coercer"
 require_relative "committee/request_unpacker"
-require_relative "committee/request_validator"
 require_relative "committee/response_generator"
 require_relative "committee/response_validator"
 require_relative "committee/router"
@@ -21,8 +20,9 @@ require_relative "committee/middleware/request_validation"
 require_relative "committee/middleware/response_validation"
 require_relative "committee/middleware/stub"
 
-require_relative "committee/schema_validator/hyper_schema"
 require_relative "committee/schema_validator/option"
+require_relative "committee/schema_validator/hyper_schema"
+require_relative "committee/schema_validator/hyper_schema/request_validator"
 
 require_relative "committee/bin/committee_stub"
